MOTECH EBODAC EBODAC-429 DEV - Fix Integration Tests in VisitControllerIT
