+++ src/test/java/org/motechproject/ebodac/utils/VisitUtils.java
@@ -2,13 +2,9 @@ package org.motechproject.ebodac.utils;
 
 
 import org.joda.time.LocalDate;
-import org.motechproject.ebodac.dao.VisitDAO;
 import org.motechproject.ebodac.domain.Subject;
 import org.motechproject.ebodac.domain.Visit;
 import org.motechproject.ebodac.domain.VisitType;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import static junit.framework.Assert.assertEquals;
 
@@ -31,16 +27,8 @@ public class VisitUtils {
         assertEquals(expected.getSubject().getSubjectId(), actual.getSubject().getSubjectId());
         assertEquals(expected.getType(), actual.getType());
         assertEquals(expected.getDate(), actual.getDate());
-        assertEquals(expected.getDateProjected(), actual.getDateProjected());
+        assertEquals(expected.getMotechProjectedDate(), actual.getMotechProjectedDate());
         assertEquals(expected.getOwner(), actual.getOwner());
     }
 
-    public static List<Visit> convertDAOVisitsToVisits(List<VisitDAO> visitsDao) {
-        ArrayList<Visit> ret = new ArrayList<Visit>();
-
-        for(VisitDAO visitDao : visitsDao) {
-            ret.add(visitDao.toVisit());
-        }
-        return ret;
-    }
 }
